Created structure to allow license to be used depending on what files are found in remotedb/license directory
Created structure to allow license to be used depending on what files
are found in remotedb/license directory. If license.crt and license.pvk
are found then the -prod response file is used, if
.BESLicenseAuthorization file is found but not the .crt and .pvk file
then the .BESLicenseAuthorization file is used - note that this is only
if license version is used and not the evaluation version.